[edit] 랜딩 페이지 및 about us 페이지 분리, 수정
