+++ client/src/App.js
@@ -14,6 +14,8 @@ import {
   Manage,
   Privacy,
   Terms,
+  OurTeam,
+  OurHistory,
   NotFound,
 } from "./pages";
 import {
@@ -98,6 +100,8 @@ function App() {
             <Route path='/donations/*' element={<Manage />} />
             <Route path='/privacy' element={<Privacy />} />
             <Route path='/terms' element={<Terms />} />
+            <Route path='/ourteam' element={<OurTeam />} />
+            <Route path='/history' element={<OurHistory />} />
             <Route path='*' element={<NotFound />} />
           </Routes>
         </div>
