+++ client/src/styles/Aboutus.css
@@ -213,7 +213,6 @@
 .about-bg-yellow {
     width: 100vw;
     height: auto;
-    margin-top: 200px;
 }
 
 .about-we-title {
