+++ client/src/styles/Home.css
@@ -3,84 +3,24 @@
 }
 
 .landing-banner {
-    margin-top: 30vh;
+    margin-top: 15vh;
     position: relative;
     height: 400px;
 }
 
 .landing-title {
-    font-size: 58px;
+    font-size: 300%;
     position: relative;
     color: #443826;
     z-index: 100;
     font-family: 'ruddy';
 }
 
-.landing-apple, 
-.landing-corn,
-.landing-radish,
-.landing-pepper,
-.landing-tomato,
-.landing-watermelon,
-.landing-paprika {
+.landing-left, 
+.landing-right {
     /* position: relative; */
     position: absolute;
     z-index: 1;
-}
-
-.landing-apple {
-    width: 11vw;
-    height: auto;
-    top: -10%;
-    left: 15%;
-    transform: translate( -50%, -50% );
-}
-
-.landing-corn {
-    width: 11vw;
-    height: auto;
-    top: 50%;
-    left: 20%;
-    transform: translate( -50%, -50% );
-}
-
-.landing-radish {
-    width: 18vw;
-    height: auto;
-    top: 40%;
-    left: 70%;
-    transform: translate( -50%, -50% );
-}
-
-.landing-tomato {
-    width: 9vw;
-    height: auto;
-    top: -20%;
-    left: 80%;
-    transform: translate( -50%, -50% );
-}
-
-.landing-pepper {
-    width: 6vw;
-    height: auto;
-    top: 20%;
-    left: 90%;
-    transform: translate( -50%, -50% );
-}
-
-.landing-watermelon {
-    width: 17vw;
-    height: auto;
-    top: 40%;
-    left: 30%;
-    transform: translate( -50%, -50% );
-}
-
-.landing-paprika {
-    width: 8vw;
-    top: -20%;
-    left: 35%;
-    transform: translate( -50%, -50% );
 }
 
 /******************* Part 1 ********************/
@@ -95,13 +35,19 @@
 }
 
 .landing-part1-title {
-    position: absolute;
-    transform: translate( -50%, -50% );
-    top: 20%;
-    left: 50%;
-    font-size: 36px;
+    font-size: 300%;
     color: #443826;
     white-space: nowrap;
+    font-family: 'ruddy';
+    line-height: 180%;  
+}
+
+.landing-part1-home-title {
+    font-size: 170%;
+    color: #443826;
+    white-space: nowrap;
+    font-family: 'ruddy';
+    line-height: 180%;
 }
 
 .landing-part1-img {
@@ -111,25 +57,25 @@
 
 .landing-part11 {
     position: absolute;
-    top: 40%;
+    top: 45%;
     left: 5%;
 }
 
 .landing-part12 {
     position: absolute;
-    top: 60%;
+    top: 65%;
     left: 27%;
 }
 
 .landing-part13 {
     position: absolute;
-    top: 40%;
+    top: 45%;
     left: 50%;
 }
 
 .landing-part14 {
     position: absolute;
-    top: 60%;
+    top: 65%;
     left: 65%;
 }
 
@@ -153,7 +99,7 @@
     transform: translate( -50%, -50% );
     padding-top: 200px;
     left: 50%;
-    font-size: 48px;
+    font-size: 300%;
     color: #443826;
     white-space: nowrap;
     font-family: 'ruddy';
@@ -203,10 +149,7 @@
 
 /******************* Part 3 ********************/
 .landing-part3 {
-    position: relative;
-    margin-top: 120px;
-    margin-bottom: 100px;
-    height: 500px;
+    margin: 15vh 0 100px;
 }
 
 .landing-part3-title {
@@ -221,15 +164,13 @@
     padding: 0 10px;
 }
 
-.landing-part3-flex {
-    display: flex;
-    justify-content: space-between;
-    margin: 0 50px;
-}
-
 .landing-part3-ment {
     font-size: 20px;
     color: #443826;
+}
+
+.landing-part3-section {
+    margin: 5vh 0;
 }
 
 /******************* Part 4 ********************/
@@ -371,11 +312,6 @@
         height: 1200px;
     }
 
-    .landing-part3-flex {
-        display: flex;
-        flex-direction: column;
-    }
-
     .landing-part3-img {
         width: 120px;
         height: 120px;
@@ -414,5 +350,27 @@
         margin: 200px 0;
         display: flex;
         position: relative;
+    }
+
+    .landing-left, .landing-right {
+        display: none;
+    }
+}
+
+@media screen and (min-width: 768px) {
+    .landing-left {
+        width: 10vw;
+        height: auto;
+        top: 35%;
+        left: 20%;
+        transform: translate( -50%, -50% );
+    }
+    
+    .landing-right {
+        width: 15vw;
+        height: auto;
+        top: 65%;
+        left: 80%;
+        transform: translate( -50%, -50% );
     }
 }
